requirements.txt
git+https://github.com/LizardM4/picamera.git@mp4
numpy>=1.13.1
python-telegram-bot>=8.0
Pillow>=4.3.0
bcrypt>=3.1.3
RPi.GPIO>=0.6.3
Pyro4>=4.73